+++ connector/ce/libs/api/src/main/kotlin/de/sovity/edc/ce/api/common/model/UiHttpOauth2PrivateKeyAuthorization.java
@@ -25,6 +25,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 
 @Data